UdaConnect/modules/location_service_grpc/requirements.txt
grpcio==1.41.1
grpcio-tools==1.41.1
kafka-python==2.2.3

